V2.0

+++ OESAppApi/Api/Controllers/UserQuizzesController.cs
@@ -1,4 +1,5 @@
-﻿using Domain.Entities.Courses;
+﻿using Domain.Entities.Common;
+using Domain.Entities.Courses;
 using Domain.Entities.Questions;
 using Domain.Entities.Quizzes;
 using Domain.Entities.UserQuizzes;
@@ -6,6 +7,7 @@ using Domain.Entities.Users;
 using Domain.Services;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.JsonPatch;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using OESAppApi.Extensions;
@@ -37,13 +39,17 @@ public class UserQuizzesController : ControllerBase
         int userId = _tokenService.GetUserId(Request.ExtractToken());
 
         UserQuiz? userQuiz = await _context.UserQuiz
-            .Where(q => q.Id == id && q.UserId == userId)
+            .Where(q => q.Id == id)
             .Include(q => q.Questions.Where(q => q.ItemId == id))
             .ThenInclude(q => q.Options)
             .SingleOrDefaultAsync();
 
         if (userQuiz is null)
             return NotFound();
+
+        var permission = await _context.UserQuizUserPermission.Where(p => p.UserId == userId && p.UserQuizId == id).SingleOrDefaultAsync();
+        if (userQuiz.UserId != userId && permission is null)
+            return Forbid(JwtBearerDefaults.AuthenticationScheme);
 
         var result = await _courseRepository.GetUserCourseRoleAsync(userQuiz.CourseId, userId);
         if (!result.IsSuccess)
@@ -59,7 +65,7 @@ public class UserQuizzesController : ControllerBase
         var result = await _courseRepository.GetUserCourseRoleAsync(courseId, userId);
 
         if (!result.IsSuccess)
-            return Forbid();
+            return Forbid(JwtBearerDefaults.AuthenticationScheme);
 
         UserQuiz q = request.ToUserQuiz(courseId, _tokenService.GetUserId(Request.ExtractToken()));
         var newQuiz = _context.Add(q);
@@ -73,9 +79,15 @@ public class UserQuizzesController : ControllerBase
     public async Task<ActionResult> Put(int id, [FromBody] UserQuizRequest value)
     {
         int userId = _tokenService.GetUserId(Request.ExtractToken());
-        UserQuiz? q = await _context.UserQuiz.SingleOrDefaultAsync(q => q.Id == id && q.UserId == userId);
+        UserQuiz? q = await _context.UserQuiz.SingleOrDefaultAsync(q => q.Id == id);
         if (q is null)
             return NotFound();
+
+        var permission = await _context.UserQuizUserPermission
+            .Where(p => p.UserId == userId && p.UserQuizId == id && p.Permission == EUserPermission.Edit)
+            .SingleOrDefaultAsync();
+        if (q.UserId != userId && permission is null)
+            return Forbid(JwtBearerDefaults.AuthenticationScheme);
 
         await _context.Question.Where(q => q.ItemId == id).ExecuteDeleteAsync();
         q.Questions = value.Questions.ToQuestionList();
@@ -93,12 +105,123 @@ public class UserQuizzesController : ControllerBase
     {
         int userId = _tokenService.GetUserId(Request.ExtractToken());
         UserQuiz? q = await _context.UserQuiz.SingleOrDefaultAsync(q => q.Id == id && q.UserId == userId);
-
-        if (q is null) return NotFound();
+        if (q is null) 
+            return NotFound();
+        
+        var permission = await _context.UserQuizUserPermission
+            .Where(p => p.UserId == userId && p.UserQuizId == id && p.Permission == EUserPermission.Edit)
+            .SingleOrDefaultAsync();
+        if (q.UserId != userId && permission is null)
+            return Forbid(JwtBearerDefaults.AuthenticationScheme);
 
         _context.UserQuiz.Remove(q);
         await _context.SaveChangesAsync();
 
         return NoContent();
+    }
+
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+    [HttpPost("{id}/permissions")]
+    public async Task<ActionResult<UserQuizUserPermissionResponse>> AddPermission(int id, [FromBody] UserQuizUserPermissionRequest request)
+    {
+        int userId = _tokenService.GetUserId(Request.ExtractToken());
+        var userQuiz = await _context.UserQuiz
+            .Include(u => u.UserPermissions)
+            .SingleOrDefaultAsync(u => u.Id == id);
+        if (userQuiz is null)
+            return NotFound();
+
+        var editorPermission = await _context.UserQuizUserPermission
+            .Where(p => p.UserQuizId == id && p.UserId == userId && p.Permission == EUserPermission.Edit)
+            .SingleOrDefaultAsync();
+        if (userQuiz.UserId != userId && editorPermission is null)
+            return Forbid(JwtBearerDefaults.AuthenticationScheme);
+
+        if (userQuiz.UserPermissions.Any(u => u.UserId == request.UserId))
+            return BadRequest();
+
+        userQuiz.UserPermissions.Add(new()
+        {
+            UserId = request.UserId,
+            Permission = request.Permission
+        });
+
+        await _context.SaveChangesAsync();
+        return NoContent();
+    }
+
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+    [HttpPut("{id}/permissions")]
+    public async Task<ActionResult<UserQuizUserPermissionResponse>> UpdatePermission(int id, [FromBody] UserQuizUserPermissionRequest request)
+    {
+        int userId = _tokenService.GetUserId(Request.ExtractToken());
+        var userQuiz = await _context.UserQuiz
+            .Include(u => u.UserPermissions)
+            .SingleOrDefaultAsync(u => u.Id == id);
+        if (userQuiz is null)
+            return NotFound();
+
+        var editorPermission = await _context.UserQuizUserPermission
+            .Where(p => p.UserQuizId == id && p.UserId == userId && p.Permission == EUserPermission.Edit)
+            .SingleOrDefaultAsync();
+        if (userQuiz.UserId != userId && editorPermission is null)
+            return Forbid(JwtBearerDefaults.AuthenticationScheme);
+
+        var permission = userQuiz.UserPermissions.SingleOrDefault(u => u.UserId == request.UserId);
+        if (permission is null)
+            return BadRequest();
+
+        permission.Permission = request.Permission;
+        await _context.SaveChangesAsync();
+        return NoContent();
+    }
+
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+    [HttpDelete("{id}/permissions/{permissionUserId}")]
+    public async Task<ActionResult<UserQuizUserPermissionResponse>> RemovePermission(int id, int permissionUserId)
+    {
+        int userId = _tokenService.GetUserId(Request.ExtractToken());
+        var userQuiz = await _context.UserQuiz
+            .Include(u => u.UserPermissions)
+            .SingleOrDefaultAsync(u => u.Id == id);
+        if (userQuiz is null)
+            return NotFound();
+        
+        var editorPermission = await _context.UserQuizUserPermission
+            .Where(p => p.UserQuizId == id && p.UserId == userId && p.Permission == EUserPermission.Edit)
+            .SingleOrDefaultAsync();
+        if (userQuiz.UserId != userId && editorPermission is null && userId != permissionUserId)
+            return Forbid(JwtBearerDefaults.AuthenticationScheme);
+
+        var permission = userQuiz.UserPermissions.SingleOrDefault(u => u.UserId == permissionUserId);
+        if (permission is null)
+            return BadRequest();
+
+        userQuiz.UserPermissions.Remove(permission);
+        await _context.SaveChangesAsync();
+        return NoContent();
+    }
+
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+    [HttpGet("{id}/permissions")]
+    public async Task<ActionResult<UserQuizUserPermissionResponse>> GetPermissions(int id)
+    {
+        int userId = _tokenService.GetUserId(Request.ExtractToken());
+        var userQuiz = await _context.UserQuiz.SingleOrDefaultAsync(u => u.Id == id);
+        if (userQuiz is null)
+            return NotFound();
+        
+        var permission = await _context.UserQuizUserPermission.Where(p => p.UserQuizId == id && p.UserId == userId).SingleOrDefaultAsync();
+        if (userQuiz.UserId != userId && permission is null)
+            return Forbid(JwtBearerDefaults.AuthenticationScheme);
+
+        List<UserQuizUserPermissionResponse> permissions = await _context.UserQuizUserPermission
+            .Include(p => p.User)
+            .Where(p => p.UserQuizId == id)
+            .Select(p => new UserQuizUserPermissionResponse(p.UserId, p.Permission,
+                p.User.FirstName, p.User.LastName, p.User.Username))
+            .ToListAsync();
+
+        return Ok(permissions);
     }
 }
